requirements.txt
ffmpeg-python
torch
torchvision
openai-whisper
torchaudio
